+++ packages/contracts-communication/test/apps/InterchainAppV1.t.sol
@@ -16,9 +16,9 @@ import {Test} from "forge-std/Test.sol";
 abstract contract InterchainAppV1Test is Test, AbstractICAppEvents, InterchainAppV1Events {
     bytes32 public constant IC_GOVERNOR_ROLE = keccak256("IC_GOVERNOR_ROLE");
 
-    uint256 public constant LOCAL_CHAIN_ID = 1337;
-    uint256 public constant REMOTE_CHAIN_ID = 7331;
-    uint256 public constant UNKNOWN_CHAIN_ID = 420;
+    uint64 public constant LOCAL_CHAIN_ID = 1337;
+    uint64 public constant REMOTE_CHAIN_ID = 7331;
+    uint64 public constant UNKNOWN_CHAIN_ID = 420;
     uint256 public constant APP_OPTIMISTIC_PERIOD = 10 minutes;
 
     IInterchainAppV1Harness public appHarness;
